+++ src/opticalc/exceptions.py
@@ -1,0 +1,13 @@
+class SpectralAveragesSummaryCalculationException(Exception):
+    """
+    Capture info about which part of integrated spectral averages
+    summary calculation process went bad.
+    """
+    pass
+
+
+class ColorCalculationException(Exception):
+    """
+    Capture info about which part of color calculation process went bad.
+    """
+    pass
